Struct iyes_perf_ui::diagnostics::PerfUiEntryMemUsage

pub struct PerfUiEntryMemUsage {
    pub label: String,
    pub color_gradient: ColorGradient,
    pub threshold_highlight: Option<f32>,
    pub smoothed: bool,
    pub precision: u8,
    pub sort_key: i32,
}

Perf UI Entry to display Bevy’s built-in Memory (RAM) Usage measurement diagnostic.

Displays the usage as a percentage.

Fields
label: String

Custom label. If empty (default), the default label will be used.

color_gradient: ColorGradient

Enable color based on value.

To disable (always use default color), set to empty ColorGradient::default().

Default: Green-Yellow-Red gradient between 25%-50%-75%.

threshold_highlight: Option<f32>

Highlight the value if above this threshold.

Default: 90%

smoothed: bool

Should we display the smoothed value or the raw value?

Default: true (smoothed)

precision: u8

Number of digits to display for the fractional (after the decimal point) part.

Default: 2

sort_key: i32

Sort Key (control where the entry will appear in the Perf UI).