Trait iyes_perf_ui::PerfUiEntry

pub trait PerfUiEntry: Component {
    type SystemParam: SystemParam + 'static;
    type Value: Debug;

    // Required methods
    fn label(&self) -> &str;
    fn update_value(
        &self,
        param: &mut <Self::SystemParam as SystemParam>::Item<'_, '_>
    ) -> Option<Self::Value>;
    fn sort_key(&self) -> i32;

    // Provided methods
    fn width_hint(&self) -> usize { ... }
    fn format_value(&self, value: &Self::Value) -> String { ... }
    fn value_color(&self, _value: &Self::Value) -> Option<Color> { ... }
    fn value_highlight(&self, _value: &Self::Value) -> bool { ... }
}

Trait for components representing entries (rows) in the Perf UI.

If you want to display your own info in Perf UI, create your own component types and implement this trait for them.

Required Associated Types

type SystemParam: SystemParam + 'static

Any system parameters you need to fetch/update the value.

type Value: Debug

The raw (unformatted) type of the value to be displayed.

Required Methods

fn label(&self) -> &str

The label text to display in the Perf UI.

fn update_value( &self, param: &mut <Self::SystemParam as SystemParam>::Item<'_, '_> ) -> Option<Self::Value>

Update the value to display in the Perf UI.

This function will be called once per frame, in the Update schedule, in the PerfUiSet::Update set.

fn sort_key(&self) -> i32

The sort key controls where the entry will appear in the Perf UI.

The recommended way to implement this is to have a field in your struct, which you can set to iyes_perf_ui::utils::next_sort_key() in your impl Default. Then return that value here.

That way, the entry will be sorted according to the order in which the user creates the entries.

Provided Methods

fn width_hint(&self) -> usize

Optional: provide a desired width for the value string.

The formatted value will be padded with spaces. This allows everything to line up nicely in the UI and prevents the UI from spontaneously resizing as the values change.

(assuming a monospace font)

fn format_value(&self, value: &Self::Value) -> String

Format the raw value into a string for display

Called every frame after update_value, unless it returned None. The value parameter is whatever that function returned.

If unimplemented, the value will be formatted with its Debug impl.

fn value_color(&self, _value: &Self::Value) -> Option<Color>

Optional: set a custom color for the value to display.

Called every frame after update_value, unless it returned None. The value parameter is whatever that function returned.

fn value_highlight(&self, _value: &Self::Value) -> bool

Optional: set whether the value should be displayed using the “highlighted” font.

Called every frame after update_value, unless it returned None. The value parameter is whatever that function returned.